+++ odoo/local-src/specific_payment_mode/__manifest__.py
@@ -18,8 +18,8 @@
     ],
     'data': [
         # Data
+        'data/depiltech_payment_mode.xml',  # Need to load this file first
         'data/account_payment_method.xml',
-        'data/depiltech_payment_mode.xml',
         'data/ir_config_parameter.xml',
         # Security
         'security/ir.model.access.csv',