+++ odoo/songs/updates/update_10_1_2/clean_translation.sql
@@ -1,0 +1,9 @@
+-- Clean bad translation
+-- which will recreate correctly when the module will be updated
+
+DELETE FROM
+    ir_translation
+WHERE
+    module = 'specific_payment_mode'
+AND lang = 'fr_FR'
+AND value = 'Mode de paiement';